+++ doc-examples/src/main/java/arez/doc/examples/at_memoize2/NetworkStatus.java
@@ -1,9 +1,11 @@
-package arez.doc.examples.at_computed;
+package arez.doc.examples.at_memoize2;
 
+import arez.ComputableValue;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
+import arez.annotations.ComputableValueRef;
+import arez.annotations.DepType;
 import arez.annotations.Memoize;
-import arez.annotations.Observable;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
 import elemental2.dom.DomGlobal;
@@ -12,20 +14,18 @@ import elemental2.dom.EventListener;
 @ArezComponent
 public abstract class NetworkStatus
 {
-  private final EventListener _listener;
-  private boolean _rawOnLine;
+  private final EventListener _listener = e -> updateOnlineStatus();
 
-  public NetworkStatus()
-  {
-    _listener = e -> updateOnlineStatus();
-    _rawOnLine = DomGlobal.navigator.onLine;
-  }
-
-  @Memoize
+  // Specify depType so can explicitly trigger a recalculation
+  // of method using reportPossiblyChanged()
+  @Memoize( depType = DepType.AREZ_OR_EXTERNAL )
   public boolean isOnLine()
   {
-    return isRawOnLine();
+    return DomGlobal.navigator.onLine;
   }
+
+  @ComputableValueRef
+  abstract ComputableValue getOnLineComputableValue();
 
   @OnActivate
   final void onOnLineActivate()
@@ -41,21 +41,10 @@ public abstract class NetworkStatus
     DomGlobal.window.removeEventListener( "offline", _listener );
   }
 
-  @Observable
-  boolean isRawOnLine()
-  {
-    return _rawOnLine;
-  }
-
-  void setRawOnLine( final boolean rawOnLine )
-  {
-    _rawOnLine = rawOnLine;
-  }
-
   @Action
   void updateOnlineStatus()
   {
-    //Updating the observable will force @Memoize method to recalculate
-    setRawOnLine( DomGlobal.navigator.onLine );
+    // Explicitly trigger a recalculation of the OnLine value
+    getOnLineComputableValue().reportPossiblyChanged();
   }
 }